Split out ember-compatibiltiy-functions with some shareable bits.
When looking to implement the ember/codegen data mode, some
functionality in ember-compatibility-functions needs sharing to
avoid extra copy&paste bloat:

- Global attribute handling via AAI split into a separate file
- Raw "data I/O" buffer split into a separate file

Moved privilege-storage and implemented a few more mock ember
methods.
