+++ src/app/util/mock/include/zap-generated/endpoint_config.h
@@ -1,2 +1,4 @@
 // Number of fixed endpoints
 #define FIXED_ENDPOINT_COUNT (3)
+
+#define ATTRIBUTE_LARGEST (1003)
